Added a cast to avoid a compiler warning.
--HG--
extra : convert_revision : svn%3Afdd8eb12-d10e-0410-9acb-85c331704f74/trunk%403813

+++ dlib/svm/svm.h
@@ -700,8 +700,8 @@ namespace dlib
             );
 
         // count the number of positive and negative examples
-        const long num_pos = sum(y > 0);
-        const long num_neg = sum(y < 0);
+        const long num_pos = (long)sum(y > 0);
+        const long num_neg = (long)sum(y < 0);
 
         // figure out how many positive and negative examples we will have in each fold
         const long num_pos_test_samples = num_pos/folds;
